@@ -2,7 +2,7 @@
 import {BrowserModule} from '@angular/platform-browser';
 import {NgModule} from '@angular/core';
 import {FormsModule} from '@angular/forms';
-import {HttpClientModule} from '@angular/common/http';
+import {provideHttpClient, withInterceptorsFromDi} from '@angular/common/http';
 import {AppComponent} from './app.component';
 import {AppRoutingModule} from './app-routing.module';
 import {ComplexPortalService} from './complex/shared/service/complex-portal.service';
@@ -14,7 +14,7 @@ import {BasketService} from './shared/basket/service/basket.service';
 import {ProgressBarComponent} from './shared/loading-indicators/progress-bar/progress-bar.component';
 import {ToastrModule} from 'ngx-toastr';
 import {LocalSearchComponent} from './search/local-search/local-search.component';
-import {NoopAnimationsModule} from '@angular/platform-browser/animations';
+import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {MatProgressBarModule} from '@angular/material/progress-bar';
 import {OntologiesComponent} from './ontologies/ontologies.component';
 import {AnalyticsService} from './shared/google-analytics/service/analytics.service';
@@ -31,6 +31,8 @@ import {NgxGoogleAnalyticsModule, NgxGoogleAnalyticsRouterModule} from 'ngx-goog
 import {environment} from '../environments/environment';
 import {FooterComponent} from './footer/footer.component';
 import {SpeciesPipe} from './complex/shared/pipe/species.pipe';
+import {ProgressSpinnerComponent} from './shared/loading-indicators/progress-spinner/progress-spinner.component';
+
 
 @NgModule({
   declarations: [
@@ -44,11 +46,10 @@ import {SpeciesPipe} from './complex/shared/pipe/species.pipe';
     FooterComponent
   ],
   imports: [
+    BrowserAnimationsModule,
     BrowserModule,
-    RouterModule.forRoot([], {relativeLinkResolution: 'legacy'}),
+    RouterModule.forRoot([], {}),
     FormsModule,
-    HttpClientModule,
-    NoopAnimationsModule,
     ToastrModule.forRoot(),
     NgxPageScrollCoreModule.forRoot({scrollOffset: 50}),
     MatProgressBarModule,
@@ -57,6 +58,7 @@ import {SpeciesPipe} from './complex/shared/pipe/species.pipe';
     NgxGoogleAnalyticsRouterModule,
     HomeModule,
     AppRoutingModule,
+    ProgressSpinnerComponent
   ],
   providers: [
     ComplexPortalService,
@@ -68,7 +70,8 @@ import {SpeciesPipe} from './complex/shared/pipe/species.pipe';
     LocalStorageService,
     SpeciesPipe,
     AnalyticsService,
-    SearchService
+    SearchService,
+    provideHttpClient(withInterceptorsFromDi())
   ],
   bootstrap: [AppComponent]
 })
