@@ -1,4 +1,4 @@
-import {Component, Input, OnInit} from '@angular/core';
+import {Component, OnInit, input } from '@angular/core';
 import {CrossReference} from '../../shared/model/complex-details/cross-reference.model';
 import {ComplexDetails} from '../../shared/model/complex-details/complex-details.model';
 
@@ -9,11 +9,9 @@ import {ComplexDetails} from '../../shared/model/complex-details/complex-details
 })
 export class ComplexEvidenceComponent implements OnInit {
 
-  private _ecoXRef: CrossReference;
-  private _intactXRefs: CrossReference[];
-  @Input()
-  complex: ComplexDetails;
-  private _flaskSymbol: string;
+  intactXRefs: CrossReference[];
+  complex = input<ComplexDetails>();
+  stars: ('empty' | 'full')[] = ['empty', 'empty', 'empty', 'empty', 'empty'];
 
 
   constructor() {
@@ -21,6 +19,11 @@ export class ComplexEvidenceComponent implements OnInit {
 
   ngOnInit() {
     this.findXRefs();
+    this.formatEvidenceOntologyStars();
+  }
+
+  private formatEvidenceOntologyStars(): void {
+    this._star(this.complex().evidenceType?.confidenceScore || 3);
   }
 
   /**
@@ -29,82 +32,23 @@ export class ComplexEvidenceComponent implements OnInit {
    * Also we add the flask symbol, which is for the icon. (Similar to the organism view)
    */
   private findXRefs(): void {
-    for (let i = 0; i < this.complex.crossReferences.length; i++) {
-      const crossRef = this.complex.crossReferences[i];
-      const database = this.complex.crossReferences[i].database;
+    for (let i = 0; i < this.complex().crossReferences.length; i++) {
+      const crossRef = this.complex().crossReferences[i];
+      const database = this.complex().crossReferences[i].database;
 
-      if (database === 'evidence ontology') {
-        this._ecoXRef = crossRef;
-        switch (this._ecoXRef.identifier) {
-          case ('ECO:0000353'):
-            this._ecoXRef.description = 'physical interaction evidence';
-            this._flaskSymbol = 'E'; // Full
-            break;
-          case ('ECO:0005610'):
-            this._ecoXRef.description = 'inferred by homology';
-            this._flaskSymbol = 'C'; // 40%
-            break;
-          case ('ECO:0005544'):
-            this._ecoXRef.description = 'inferred by orthology';
-            this._flaskSymbol = 'C'; // 40%
-            break;
-          case ('ECO:0005546'):
-            this._ecoXRef.description = 'inferred by paralogy';
-            this._flaskSymbol = 'C'; // 40%
-            break;
-          case ('ECO:0005547'):
-            this._ecoXRef.description = 'inferred by curator';
-            this._flaskSymbol = 'B'; // Empty
-            break;
-          case ('ECO:0005543'):
-            this._ecoXRef.description = 'inferred from mixed species evidence';
-            this._flaskSymbol = 'E'; // Full
-            break;
-          case ('ECO:0005542'):
-            this._ecoXRef.description = 'inferred from single species evidence';
-            this._flaskSymbol = 'E'; // Full
-            break;
-          case ('ECO:0007653'):
-            this._ecoXRef.description = 'machine-learning prediction';
-            this._flaskSymbol = '';
-            // TODO remove this setter once the API support predicted field
-            this.complex.predicted = true;
-            break;
-          case ('ECO:0008006'):
-            this._ecoXRef.description = 'deep-learning prediction';
-            this._flaskSymbol = '';
-            // TODO remove this setter once the API support predicted field
-            this.complex.predicted = true;
-            break;
-        }
-      }
       if (database === 'intact') {
-        if (this._intactXRefs === undefined) {
-          this._intactXRefs = [];
+        if (this.intactXRefs === undefined) {
+          this.intactXRefs = [];
         }
-        this._intactXRefs.push(crossRef);
+        this.intactXRefs.push(crossRef);
       }
     }
   }
 
-
-  get ecoXRef(): CrossReference {
-    return this._ecoXRef;
-  }
-
-  set ecoXRef(value: CrossReference) {
-    this._ecoXRef = value;
-  }
-
-  get intactXRefs(): CrossReference[] {
-    return this._intactXRefs;
-  }
-
-  set intactXRefs(value: CrossReference[]) {
-    this._intactXRefs = value;
-  }
-
-  get flaskSymbol(): string {
-    return this._flaskSymbol;
+  _star(amount: number) {
+    this.stars.fill('full');
+    if (amount < this.stars.length) {
+      this.stars.fill('empty', amount);
+    }
   }
 }
