@@ -1,10 +1,10 @@
 import {catchError} from 'rxjs/operators';
 import {Injectable} from '@angular/core';
-import {HttpClient, HttpErrorResponse} from '@angular/common/http';
+import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 
 
 import {environment} from '../../../../../../environments/environment';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {throwError} from 'rxjs/internal/observable/throwError';
 
 const baseURL = environment.reactome_base_url;
