@@ -1,4 +1,4 @@
-import {Component, EventEmitter, Input, Output} from '@angular/core';
+import {Component, output, input } from '@angular/core';
 import {ComplexSearchResult} from '../../shared/model/complex-results/complex-search.model';
 import {Interactor} from '../../shared/model/complex-results/interactor.model';
 
@@ -9,14 +9,14 @@ import {Interactor} from '../../shared/model/complex-results/interactor.model';
 })
 
 export class ComplexNavigatorComponent {
-  @Input() complexSearch: ComplexSearchResult;
-  @Input() interactors: Interactor[];
-  @Input() canAddComplexesToBasket: boolean;
-  @Input() canRemoveComplexesFromBasket: boolean;
-  @Output() onComplexRemovedFromBasket: EventEmitter<string> = new EventEmitter<string>();
+  complexSearch = input<ComplexSearchResult>();
+  interactors = input<Interactor[]>();
+  canAddComplexesToBasket = input<boolean>();
+  canRemoveComplexesFromBasket = input<boolean>();
+  onComplexRemovedFromBasket = output<string>();
 
   interactorsSorting = 'Occurrence';
   organismIconDisplay = true;
   interactorTypeDisplay = true;
-  IDDisplay = true;
+  idDisplay = true;
 }
