@@ -7,15 +7,15 @@ import { SpeciesPipe } from './shared/pipe/species.pipe';
   imports: [RouterModule.forChild([
     {
       path: 'organisms',
-      loadChildren: () => import('app/complex/complex-organisms/complex-organisms.module').then(m => m.ComplexOrganismsModule)
+      loadChildren: () => import('./complex-organisms/complex-organisms.module').then(m => m.ComplexOrganismsModule)
     },
     {
       path: 'search',
-      loadChildren: () => import('app/complex/complex-results/complex-results.module').then(m => m.ComplexResultsModule)
+      loadChildren: () => import('./complex-results/complex-results.module').then(m => m.ComplexResultsModule)
     },
     {
       path: ':id',
-      loadChildren: () => import('app/complex/complex-details/complex-details.module').then(m => m.ComplexDetailsModule)
+      loadChildren: () => import('./complex-details/complex-details.module').then(m => m.ComplexDetailsModule)
     },
   ]),
     CommonModule,
