@@ -6,5 +6,5 @@ export interface Element {
   organismName: string;
   description: string;
   interactors: ComplexComponent[];
-  predicted?: boolean;
+  predictedComplex?: boolean;
 }
