@@ -3,7 +3,9 @@ import {Facet} from './facet.model';
  * Created by mkoch on 31/10/2016.
  */
 export interface Facets {
-  ptype_f: Facet;
-  species_f: Facet;
-  pbiorole_f: Facet;
+  ptype_f: Facet[];
+  species_f: Facet[];
+  pbiorole_f: Facet[];
+  predicted_complex_f: Facet[];
+  confidence_score_f: Facet[];
 }
