@@ -4,12 +4,10 @@ import {Injectable} from '@angular/core';
 import {HttpClient, HttpErrorResponse, HttpParams} from '@angular/common/http';
 
 import {environment} from '../../../../environments/environment';
-
+import {Observable, throwError} from 'rxjs';
 
 import {ComplexDetails} from '../model/complex-details/complex-details.model';
 import {ComplexSearchResult} from '../model/complex-results/complex-search.model';
-import {Observable} from 'rxjs/Observable';
-import {throwError} from 'rxjs/internal/observable/throwError';
 import {Element} from '../model/complex-results/element.model';
 import {Facet} from '../model/complex-results/facet.model';
 
@@ -17,6 +15,20 @@ const baseURL = environment.complex_ws_base_url;
 
 @Injectable()
 export class ComplexPortalService {
+
+  private static SPECIES_FACET_FIELD = 'species_f';
+  private static COMPONENT_TYPE_FACET_FIELD = 'ptype_f';
+  private static BIO_ROLE_FACET_FIELD = 'pbiorole_f';
+  private static PREDICTED_FACET_FIELD = 'predicted_complex_f';
+  private static CONFIDENCE_SCORE_FACET_FIELD = 'confidence_score_f';
+
+  private static FACETS = [
+    ComplexPortalService.SPECIES_FACET_FIELD,
+    ComplexPortalService.COMPONENT_TYPE_FACET_FIELD,
+    ComplexPortalService.BIO_ROLE_FACET_FIELD,
+    ComplexPortalService.PREDICTED_FACET_FIELD,
+    ComplexPortalService.CONFIDENCE_SCORE_FACET_FIELD
+  ].join(',');
 
   constructor(private http: HttpClient) {
   }
@@ -28,12 +40,7 @@ export class ComplexPortalService {
    */
   getComplex(ac: string): Observable<ComplexDetails> {
     const url = `${baseURL}/details/${ac}`;
-    // TODO Remove random predicted when real predicted complexes available
     return this.http.get<ComplexDetails>(url).pipe(
-      // map(data => {
-      //   data.predicted = Math.random() < 0.5;
-      //   return data;
-      // }),
       catchError(this.handleError)
     );
   }
@@ -45,12 +52,7 @@ export class ComplexPortalService {
    */
   getComplexAc(complexAc: string): Observable<ComplexDetails> {
     const url = `${baseURL}/complex/${complexAc}`;
-    // TODO Remove random predicted when real predicted complexes available
     return this.http.get<ComplexDetails>(url).pipe(
-      // map(data => {
-      //   data.predicted = Math.random() < 0.5;
-      //   return data;
-      // }),
       catchError(this.handleError)
     );
   }
@@ -75,31 +77,51 @@ export class ComplexPortalService {
     return this.http.get(baseURL + '/export/' + ac).pipe(catchError(this.handleError));
   }
 
+  findComplexNoFilters(query: string,
+                       currentPageIndex = 1,
+                       pageSize = 10,
+                       format = 'json'): Observable<ComplexSearchResult> {
+    return this.findComplex(query, [], [], [], [], [], currentPageIndex, pageSize, format);
+  }
+
   /**
    * Find a complex based on indexed term
    * @param query
    * @param speciesFilter
    * @param bioRoleFilter
    * @param interactorTypeFilter
+   * @param predictedFilter
+   * @param confidenceScoreFilter
    * @param currentPageIndex
    * @param pageSize
    * @param format
-   * @param facets
    * @returns {Observable<ComplexSearchResult>}
    */
-  findComplex(query: string, speciesFilter: string[] = [], bioRoleFilter: string[] = [],
-              interactorTypeFilter: string[] = [], currentPageIndex = 1, pageSize = 10,
-              format = 'json', facets = 'species_f,ptype_f,pbiorole_f'): Observable<ComplexSearchResult> {
+  findComplex(query: string,
+              speciesFilter: string[] = [],
+              bioRoleFilter: string[] = [],
+              interactorTypeFilter: string[] = [],
+              predictedFilter: string[] = [],
+              confidenceScoreFilter: string[] = [],
+              currentPageIndex = 1,
+              pageSize = 10,
+              format = 'json'): Observable<ComplexSearchResult> {
 
     let filters = '';
     if (speciesFilter.length !== 0) {
-      filters += 'species_f:(' + '"' + speciesFilter.join('"AND"') + '"' + '),';
+      filters += this.buildFilterParam(ComplexPortalService.SPECIES_FACET_FIELD, speciesFilter);
     }
     if (bioRoleFilter.length !== 0) {
-      filters += 'pbiorole_f:(' + '"' + bioRoleFilter.join('"AND"') + '"' + '),';
+      filters += this.buildFilterParam(ComplexPortalService.BIO_ROLE_FACET_FIELD, bioRoleFilter);
     }
     if (interactorTypeFilter.length !== 0) {
-      filters += 'ptype_f:(' + '"' + interactorTypeFilter.join('"AND"') + '"' + '),';
+      filters += this.buildFilterParam(ComplexPortalService.COMPONENT_TYPE_FACET_FIELD, interactorTypeFilter);
+    }
+    if (predictedFilter.length !== 0) {
+      filters += this.buildFilterParam(ComplexPortalService.PREDICTED_FACET_FIELD, predictedFilter);
+    }
+    if (confidenceScoreFilter.length !== 0) {
+      filters += this.buildFilterParam(ComplexPortalService.CONFIDENCE_SCORE_FACET_FIELD, confidenceScoreFilter);
     }
 
     /** HttpParams is immutable. Its set() method returns a new HttpParams, without mutating the original one **/
@@ -107,17 +129,15 @@ export class ComplexPortalService {
       .set('first', ((currentPageIndex * pageSize) - pageSize).toString())
       .set('number', pageSize.toString())
       .set('format', format)
-      .set('facets', facets)
+      .set('facets', ComplexPortalService.FACETS)
       .set('filters', filters);
 
-    // TODO Remove random predicted when real predicted complexes available
     return this.http.get<ComplexSearchResult>(baseURL + '/search/' + query, {params: params}).pipe(
-      // map(result => {
-          // result.elements.forEach(e => e.predicted = Math.random() < 0.5);
-          // return result;
-        // }
-      // ),
       catchError(this.handleError));
+  }
+
+  private buildFilterParam(filterField: string, filterValues: string[]): string {
+    return filterField + ':(' + '"' + filterValues.join('"OR"') + '"' + '),';
   }
 
   private handleError(err: HttpErrorResponse | any): Observable<any> {
@@ -129,10 +149,7 @@ export class ComplexPortalService {
   }
 
   getSimplifiedComplex(complexAc: string): Observable<Element> {
-    const url = `${baseURL}/complex-simplified/${complexAc}`;
-    // const url = `${baseURL}/complex-simplified/${complexAc}`;
-    return this.http.get(url).pipe(
-      catchError(this.handleError));
+    return this.http.get(`${baseURL}/complex-simplified/${complexAc}`).pipe(catchError(this.handleError));
   }
 
 }
