@@ -7,9 +7,9 @@ const EBI_BASE_URL = 'https://www.ebi.ac.uk/';
 const COMPLEX_BASE_URL = 'https://wwwdev.ebi.ac.uk/';
 const LICENSE_URL = 'https://creativecommons.org/licenses/by/4.0/';
 const REACTOME_BASE_URL = 'https://reactome.org';
-const INTACT_FTP_BASE_URL = 'ftp://ftp.ebi.ac.uk/pub/databases/intact/';
+const INTACT_FTP_BASE_URL = 'https://ftp.ebi.ac.uk/pub/databases/intact/';
 const COMPLEX__FTP_BASE_URL = INTACT_FTP_BASE_URL + 'complex/';
-const COMPLEX_PORTAL_DOCUMENTATION_URL = 'https://raw.githubusercontent.com/Complex-Portal/complex-portal-documentation/master/';
+const COMPLEX_PORTAL_DOCUMENTATION_URL = 'https://raw.githubusercontent.com/Complex-Portal/complex-portal-documentation/uniplex-import/';
 const INTACT_PORTAL_DOCUMENTATION_URL = 'https://raw.githubusercontent.com/Intact-Portal/intact-portal-documentation/master/';
 
 export const environment = {
